Fix fuzzy_compare to properly work with arrays.
Actually this is a rewrite of fuzzy_compare which primarily make it work
with arrays.

Current implementation was not correct with more than one element in the
array. What is more, I believe that when comparing kubernetes
descriptors we should compare arrays by their content *ignoring element
order*. That's because kubernetes will reorder elements in arrays (I
can't think of a case where element order is significant) after initial
descriptor upload.