Fix #25,#27,#28 - remove Kind/apiVersion from kubernetes types
Using the code in PR garethr/puppet-swagger-generator#9, the providers
and types have been rebuilt without the `apiVersion` and `kind`
properties that have no use since they are generated by kubeclient.
It also removes the `stringData` from kubernetes_secret because
this is a write-only property which can't work in Puppet.
This also renames any Puppet property in those types that are camel-
cased to snake case internally, so that puppet bug PUP-7140 doesn't
trigger.

Signed-off-by: Brice Figureau <brice@daysofwonder.com>