Fix collection names
Kubeclient and kubernetes assumes that collections of objects are
the object name plural. We use to just add an 's' at the end of the
object name to produce plural.
Except this really doesn't work for objects that already end with an
's', like 'ingress'.
Let's be a bit smarter :)